moved case-insensitive find function to parameter tree
