+++ src/rig_reconfigure.cpp
@@ -33,7 +33,6 @@ static void glfw_error_callback(int error, const char *description) {
 
 void visualizeParameters(ServiceWrapper &serviceWrapper, const std::shared_ptr<ParameterGroup> &parameterNode,
                          std::size_t maxParamLength, const std::string &filterString, const std::string &prefix = "");
-std::size_t findCaseInsensitive(const std::string &string, const std::string &pattern);
 void highlightedText(const std::string &text, std::size_t start, std::size_t end,
                      const ImVec4 highlightColor = FILTER_HIGHLIGHTING_COLOR);
 
@@ -407,15 +406,6 @@ void visualizeParameters(ServiceWrapper &serviceWrapper, const std::shared_ptr<P
             }
         }
     }
-}
-
-// based on the following stack overflow answer: https://stackoverflow.com/a/19839371
-std::size_t findCaseInsensitive(const std::string &string, const std::string &pattern) {
-    auto it = std::search(string.begin(), string.end(), pattern.begin(), pattern.end(),
-            [](unsigned char ch1, unsigned char ch2) { return std::toupper(ch1) == std::toupper(ch2); }
-    );
-
-    return (it != string.end()) ? std::distance(string.begin(), it) : std::string::npos;
 }
 
 void highlightedText(const std::string &text, std::size_t start, std::size_t end,
